Revert "Configure the database and restructure the directory for multi apps"
