+++ src/routes/user.route.js
@@ -5,4 +5,3 @@ import { userController } from '../controllers';
 export default express
 	.Router()
 	.get('/', userController.getAllUsers)
-	.post('/', userController.createUser);
